+++ engine/src/execution_graph/logic_controllers/taskflow/graph.cpp
@@ -97,7 +97,7 @@ namespace cache {
 			}
 		}
 		// Lets iterate through the futures to check for exceptions
-		for(int i = 0; i < futures.size(); i++){
+		for(size_t i = 0; i < futures.size(); i++){
 			futures[i].get();
 		}
