+++ engine/src/utilities/DebuggingUtils.cpp
@@ -46,7 +46,7 @@ void print_blazing_table_view(ral::frame::BlazingTableView table_view, const std
 	std::cout<<"Table: "<<table_name<<std::endl;
 	std::cout<<"\t"<<"Num Rows: "<<table_view.num_rows()<<std::endl;
 	std::cout<<"\t"<<"Num Columns: "<<table_view.num_columns()<<std::endl;
-	for(size_t col_idx=0; col_idx<table_view.num_columns(); col_idx++){
+	for(int col_idx=0; col_idx<table_view.num_columns(); col_idx++){
 		std::string col_string;
 		if (table_view.num_rows() > 0){
 			col_string = cudf::test::to_string(table_view.column(col_idx), "|");
@@ -64,7 +64,7 @@ std::string blazing_table_view_schema_to_string(ral::frame::BlazingTableView tab
 	ostream <<"Table: "<<table_name<<std::endl;
 	ostream<<"\t"<<"Num Rows: "<<table_view.num_rows()<<std::endl;
 	ostream<<"\t"<<"Num Columns: "<<table_view.num_columns()<<std::endl;
-	for(size_t col_idx=0; col_idx<table_view.num_columns(); col_idx++){
+	for(int col_idx=0; col_idx<table_view.num_columns(); col_idx++){
 		ostream<<"\t"<<table_view.names().at(col_idx)<<" ("<<"type: "<<type_string(table_view.column(col_idx).type())<<")"<<std::endl;
 	}
 	return ostream.str();
